Changed the ints to bytes to make Tyler happy

+++ src/TileType.java
@@ -1,21 +1,21 @@
 
 public abstract class TileType {
-	static final int GRASS = 0;
-	static final int GRASS_FLOWER = 1;
-	static final int GRASS_FLOWER1 = 2;
-	static final int GRASS_FLOWER2 = 3;
-	static final int TREE = 4;
-	static final int TREE_BARE = 5;
-	static final int WALL = 6;
-	static final int RIGHT_WALL = 7;
-	static final int LEFT_WALL = 8;
-	static final int INTERIOR_WALL = 9;
-	static final int RIGHT_COLUMN = 10;
-	static final int LEFT_COLUMN = 11;
-	static final int STATUE = 12;
-	static final int DOOR = 13;
-	static final int INTERIOR = 14;
-	static final int ROCK = 15;
-	static final int DIRT = 16;
-	static final int MAP_ERR = -1;
+	static final byte GRASS = 0;
+	static final byte GRASS_FLOWER = 1;
+	static final byte GRASS_FLOWER1 = 2;
+	static final byte GRASS_FLOWER2 = 3;
+	static final byte TREE = 4;
+	static final byte TREE_BARE = 5;
+	static final byte WALL = 6;
+	static final byte RIGHT_WALL = 7;
+	static final byte LEFT_WALL = 8;
+	static final byte INTERIOR_WALL = 9;
+	static final byte RIGHT_COLUMN = 10;
+	static final byte LEFT_COLUMN = 11;
+	static final byte STATUE = 12;
+	static final byte DOOR = 13;
+	static final byte INTERIOR = 14;
+	static final byte ROCK = 15;
+	static final byte DIRT = 16;
+	static final byte MAP_ERR = -1;
 }
